Hades_Lite/Hades Web/hades_api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
PyJWT>=2.8.0
cryptography>=41.0.0
redis>=5.0.0
celery>=5.3.0
